Fiks feilende cypress-test
Feilen kom av at vi hadde en veldig rar avhengighetsgraf som gjorde at test-filen
til slutt forsøkte å importere komponent-CSS inn i selve testen.

Fikser det per nå ved å flytte relevante funksjoner til middleware/api.ts (i stedet for at de ligger i ducks/utils.ts).
Dette stanser avhengighetsgrafen her slik at testen ikke forsøker å importere CSS.

> Error: Webpack Compilation Error
> ./src/components/modal/veiledergruppe/veiledergruppe-modal.css 1:0
> Module parse failed: Unexpected character '#' (1:0)
> You may need an appropriate loader to handle this file type, currently no loaders are configured to process this file. See https://webpack.js.org/concepts#loaders
> > #veiledergruppe-modal__valgteveileder__label {
> |     margin-bottom: 0.5rem;
> | }
>  @ ./src/components/modal/veiledergruppe/veiledergruppeendring-feilet-modal.tsx 3:0-36
>  @ ./src/components/modal/modal-enhet-side-controller.tsx
>  @ ./src/enhetsportefolje/enhet-side.tsx
>  @ ./src/routes.tsx
>  @ ./src/application.tsx
>  @ ./src/ducks/ui/listevisning-selectors.ts
>  @ ./src/ducks/ui/listevisning.ts
>  @ ./src/ducks/utils.ts
>  @ ./src/middleware/api.ts
>  @ ./src/utils/utils.tsx
>  @ ./cypress/e2e/sidebar_spec.js
